removed db.py
