+++ ckanext/datastore/tests/test_helpers.py
@@ -7,7 +7,7 @@ from ckan.common import config
 import ckanext.datastore.helpers as datastore_helpers
 import ckanext.datastore.backend.postgres as postgres_backend
 import ckanext.datastore.tests.helpers as datastore_test_helpers
-import ckanext.datastore.db as db
+import ckanext.datastore.backend.postgres as db
 
 
 eq_ = nose.tools.eq_
@@ -71,9 +71,7 @@ class TestGetTables(object):
         if not config.get('ckan.datastore.read_url'):
             raise nose.SkipTest('Datastore runs on legacy mode, skipping...')
 
-        engine = db._get_engine(
-            {'connection_url': config['ckan.datastore.write_url']}
-        )
+        engine = db._get_engine(config['ckan.datastore.write_url'])
         cls.Session = orm.scoped_session(orm.sessionmaker(bind=engine))
 
         datastore_test_helpers.clear_db(cls.Session)
