+++ ckanext/datastore/tests/test_info.py
@@ -12,7 +12,7 @@ import ckan.lib.create_test_data as ctd
 import ckan.model as model
 from ckan.tests.legacy import is_datastore_supported
 
-import ckanext.datastore.db as db
+import ckanext.datastore.backend.postgres as db
 from ckanext.datastore.tests.helpers import extract, rebuild_all_dbs
 
 import ckan.tests.helpers as helpers
