+++ ckanext/datastore/tests/test_unit.py
@@ -7,11 +7,13 @@ import mock
 from ckan.common import config
 
 import ckan.tests.legacy as tests
-import ckanext.datastore.db as db
+import ckanext.datastore.backend.postgres as db
 import ckanext.datastore.helpers as helpers
 import ckanext.datastore.plugin as plugin
 import ckanext.datastore.backend.postgres as backend
 
+postgres_backend = backend.DatastorePostgresqlBackend()
+postgres_backend.configure(config)
 
 class TestTypeGetters(unittest.TestCase):
     def test_is_valid_field_name(self):
@@ -37,8 +39,7 @@ class TestTypeGetters(unittest.TestCase):
     def test_pg_version_check(self):
         if not tests.is_datastore_supported():
             raise nose.SkipTest("Datastore not supported")
-        engine = db._get_engine(
-            {'connection_url': config['sqlalchemy.url']})
+        engine = db._get_engine(config['sqlalchemy.url'])
         connection = engine.connect()
         assert db._pg_version_is_at_least(connection, '8.0')
         assert not db._pg_version_is_at_least(connection, '10.0')
@@ -46,18 +47,16 @@ class TestTypeGetters(unittest.TestCase):
 
 class TestLegacyModeSetting():
 
-    @mock.patch('ckanext.datastore.db._pg_version_is_at_least')
+    @mock.patch('ckanext.datastore.backend.postgres._pg_version_is_at_least')
     def test_legacy_mode_set_if_no_read_url_and_pg_9(self, pgv):
-
         pgv.return_value = True
 
         test_config = {
             'ckan.datastore.write_url': config['ckan.datastore.write_url'],
         }
+        assert postgres_backend.is_legacy_mode(test_config)
 
-        assert backend._is_legacy_mode(test_config)
-
-    @mock.patch('ckanext.datastore.db._pg_version_is_at_least')
+    @mock.patch('ckanext.datastore.backend.postgres._pg_version_is_at_least')
     def test_legacy_mode_set_if_no_read_url_and_pg_8(self, pgv):
 
         pgv.return_value = False
@@ -65,10 +64,9 @@ class TestLegacyModeSetting():
         test_config = {
             'ckan.datastore.write_url': config['ckan.datastore.write_url'],
         }
+        assert postgres_backend.is_legacy_mode(test_config)
 
-        assert backend._is_legacy_mode(test_config)
-
-    @mock.patch('ckanext.datastore.db._pg_version_is_at_least')
+    @mock.patch('ckanext.datastore.backend.postgres._pg_version_is_at_least')
     def test_legacy_mode_set_if_read_url_and_pg_8(self, pgv):
 
         pgv.return_value = False
@@ -77,10 +75,10 @@ class TestLegacyModeSetting():
             'ckan.datastore.write_url': config['ckan.datastore.write_url'],
             'ckan.datastore.read_url': 'some_test_read_url',
         }
+        assert postgres_backend.is_legacy_mode(test_config)
 
-        assert backend._is_legacy_mode(test_config)
 
-    @mock.patch('ckanext.datastore.db._pg_version_is_at_least')
+    @mock.patch('ckanext.datastore.backend.postgres._pg_version_is_at_least')
     def test_legacy_mode_not_set_if_read_url_and_pg_9(self, pgv):
 
         pgv.return_value = True
@@ -90,4 +88,4 @@ class TestLegacyModeSetting():
             'ckan.datastore.read_url': 'some_test_read_url',
         }
 
-        assert not backend._is_legacy_mode(test_config)
+        assert not postgres_backend.is_legacy_mode(test_config)
